Support testing with Jest; utilize Webpack more; prepare for jQuery removal
