+++ .babelrc
@@ -1,9 +1,3 @@
 {
-  "presets": [
-    ["env", {
-      "es2015": {
-        "modules": false
-      }
-    }]
-  ]
+  "presets": ["@babel/preset-env"]
 }
