+++ modules/lazyload-all/index.js
@@ -1,4 +1,9 @@
 import './styles.scss';
 
-import '../lazyload-youtube';
-import '../lazyload-vimeo';
+import '../lazyload-vimeo/lazyloadVimeo';
+
+import lazyloadVimeo from '../lazyload-vimeo';
+import lazyloadYoutube from '../lazyload-youtube';
+
+lazyloadVimeo();
+lazyloadYoutube();
