+++ modules/utils/onReady.js
@@ -1,0 +1,7 @@
+export default function onReady(fn) {
+  if (document.attachEvent ? document.readyState === 'complete' : document.readyState !== 'loading') {
+    fn();
+  } else {
+    document.addEventListener('DOMContentLoaded', fn);
+  }
+}
